Consolidated RuCl scripts to one with flag for mode, added changes to RuCl notebook.
